added new python file to contain SammyFitConfig class.
