+++ .gitignore
@@ -190,3 +190,6 @@ cython_debug/
 *.cov
 *.COV
 *.ini
+
+# Ignore mac stuff.
+*.DS_Store
