+++ examples/configFiles/uranium.ini
@@ -1,6 +1,6 @@
 [main]
-sammy_run_method = "compiled"
-sammy_command = "sammy"
+sammy_run_method = "docker"
+sammy_command = "sammy-docker"
 run_with_endf = false
 fit_energy_min = 1
 fit_energy_max = 100.0
