+++ examples/scripts/makeCompoundFit.py
@@ -4,6 +4,8 @@ from pleiades import sammyUtils, sammyOutput, sammyPlotter
 # Load the configuration file from the ini file in the parent directory
 uranium = sammyUtils.SammyFitConfig('../configFiles/uranium.ini')
 
+print(uranium.params['sammy_run_method'])
+
 # Create the needed parFiles from ENDF for the isotopes in the configuration file
 sammyUtils.create_parFile_from_endf(uranium,verbose_level=1)
